iOS 4.0 task completion now used to finish updating stops in the background if the user moves the app to the background before an update is finished.
Also, when moving to the foreground, the stop view controller reloads the (old) data and displays extrapolated times for the stops. i.e., if OBA says a bus is coming in 8 minutes and the user moves the app to the background and comes back in 5 minutes, OBA will immediately display "3 minutes" while a new update is triggered.

Added UIDevice additions which will tell us if the device/OS supports multitasking in a way that's compatible in iPhone OS 3.


Copied from https://onebusaway-iphone.googlecode.com/svn/trunk/, rev. 48 by rizzuh @ 7/22/2010 8:55 AM