State is now always restored, and we no longer default to another view after a given amount of time.
If the application is subsequently terminated (while in the background, suspended), the app gets no notification, so we need to save state when entering the background. Devices running iOS 3.x will receive the "terminate" message which will also save state.

Restoring the last state is the most common iPhone UX with apps. This is even more true with iOS 4.0, which by default will "restore state" transparently when the app returns from the foreground. If the app were to sometimes enter a restored state, but other times move to the Map screen, based on on the whims of the OS closing terminating OBA process then our users would have an inconsistent experience.

The preference for the default screen is now ignored and will be removed in a subsequent check-in.

Another change was initializing the Ideascale widget when accessing the "Feedback" tab, so that way the widget has no affect on application start up time for the vast majority of runs which will never interface with the Ideascale widget.
Copied from https://onebusaway-iphone.googlecode.com/svn/trunk/, rev. 57 by rizzuh @ 7/26/2010 2:12 AM