Removing Settings.bundle and code references to user prefs.
With the removal of a default startup screen, the settings bundle only exposed the API server and a "clear cache" setting. Both are low-level, rare use-cases th
at we may not want to support with a Settings.bundle. If necessary, it would be better to implement these in-app. (I'm not sure if it's necessary to have either
 setting.)
Copied from https://onebusaway-iphone.googlecode.com/svn/trunk/, rev. 58 by rizzuh @ 7/26/2010 2:58 AM