Fixed bug with annotations and with visible region after searching for a route.
_mapAnnotations seemed to not be too reliable and sometimes I'd see routes overlap
(i.e. the 8 stops wouldn't be cleared after searching for route 43). I got rid of that array; 
we now use [_mapView annotations] directly. This may need to be hacked
if we move to using the location beacon as I want to.

If we have stops that are close to us, we put the center of the map on the current location
as in the earlier implementation. For routes where there are no stops around us, we zoom
out to fully encompass the route and ignore the current location unlike the earlier
implementation.

Also changed the nearby stops button to be an verb-command style deal like the other
options ("see nearby stops" vs just "nearby stops").
Copied from https://onebusaway-iphone.googlecode.com/svn/trunk/, rev. 78 by rizzuh @ 8/18/2010 7:55 AM