Added the Uncrustify config
To use it, install BBUncrustifyPlugin. It will automatically detect
the config file and allow quick cleanup of selected lines, the active file,
or even a selection of files. If something seems inconsistent, install
UncrustifyX, and use the Edit Uncrustify Configuration option to tweak
the config. All of these options are in the XCode Edit menu.

Also updated the Contributing Guidelines and README to indicate our use
of the New York Times Style Guide and Uncrustify.