Show correct Trip Details times for 'scheduled arrival' buses. Fixes #359
It appears that for scheduled arrivals—I guess not surprisingly—there's no trip status object on the tripDetails object in OBATripScheduleListViewController.m (specifically relevant on line 313 of that file). When _tripDetails.status is nil, the date that is loaded into the schedule list view controller is actually in 1970, not 2015. When status is represented by an object, the date is correctly in 2015.

This commit fixes that issue by properly setting a base date (now) for cases where there isn't a status object. Also adds some debug info to OBATripStopTimeV2.