Fixes #413 — Blurry pins on iOS 8.3+
The stop icon factory was being instantiated *way* too early in the application's lifecycle. Loading images too early caused the UIScreen's scale factor to not be properly set (it was reporting as 1.0 on a retina device). In general we shouldn't really be overriding the application delegate's init method, as this is run before the underlying infrastructure is even spun up (run loop, window, etc.). There must've been a change in the UIScreen initialization sequence in iOS 8.3.

Additionally, the OBAStopIconFactory was unnecessarily caching the UIImage objects for each icon. [UIImage imageNamed:] is already cached by the system.