Move UI creation into a separate class
I want to be able to start iterating aggressively on the OBA user experience sometime soon. However, I don't want to do this at the expense of stopping releases or indefinitely maintaining a separate 'beta' branch. Moreover, it is critical to the success of any major Ux overhaul to have the users bought in on it. To that end, I want to make sure that end users have the opportunity to use a new user interface before it's completely finished.

Additionally, proper white label support for OBA will require providing a toolkit to transit agencies for them to be able to more easily design their overall application experience. These changes start moving us in that direction.

Related to:

* #375 - Support White Labeling OBA

Changes:

* Break out the UI creation aspects of OBA's application startup into a separate class
* Reverse naming of opt-in/opt-out for analytics tracking to clarify usage