Make long stop names wrap properly on the Arrival and Departure view
Fixes #691 - Arrival and Departure UI needs polishing

It turns out that there are bugs with handling long text in a table cell with the Value1 style. By using our own custom cell and wrapping replacement labels in that cell with a stack view, we can properly handle displaying long strings.