Be less cavalier about replacing table rows inside of blocks on Bookmarks
Fixes #689 - Occasional crash when reloading the bookmarks page

Instead of simply replacing and reloading an index path that may no longer exist on the bookmark controller's table view (because the user toggled a group), check to see if the row exists, and then reload it. If it doesn't exist, then reload the whole table.