Stop tearing down the reachability object on region changes
Fixes #767 - Reachability causing crashes (https://github.com/OneBusAway/onebusaway-iphone/issues/767)

The only crashes we've seen thus far in OBA 2.6.0 are happening deep within the bowels of the Reachability object. My supposition on what's happening is that when the reachability object is torn down in -[OBAApplication restartReachability] it is not releasing all of its resources/callbacks/whatever. And when one of these finally calls back after the reachability object has been torn down, the app crashes.

So, we will drop the reachability check for the user's current region server and, instead, just make sure that the Internet is accessible. 99% as good, and hopefully 100% fewer crashes.