Prospective fix for missing map annotations
Hopefully fixes #773 - Map doesn't consistently show stops when it first loads (https://github.com/OneBusAway/onebusaway-iphone/issues/773)

It appears that what was causing this issue to manifest was the thrashing of the current region on app launch. I've improved the -isEqual: method for regions and stopped updating the region if the current region is equal to the parameter. Also, I eliminated an assert that was triggering on a perfectly reasonable condition.