Update app version to 17.1.0
I think version numbers are kind of lame. Chrome and Firefox release a new 'major version' every month-and-a-half, but Apple hasn't released a new major version of their desktop operating system since 2001.

OneBusAway has been on version 2.something since 2013. Since then, the application has been virtually rewritten. But, I still haven't changed the major version number because I was saving '3.0' for something special. I have since decided that this is silly.

Instead of tying major and minor version numbers to particular features, OBA for iOS will—henceforth—treat the version number as nothing more than an analog for the date.

17.1.0:

Year: (20)17
Month: 1 (January)
Release: 0 (zero indexed)

If a bug fix release comes out in January 2017, it will be 17.1.1. If a bug fix release comes out in February, it will be 17.2.0. If a major feature release comes out in February after that, it will be 17.2.1.

I hope to be less hung up on the version number and more hung up on delivering awesome software to our users. <3.