Rearrange OBANavigationTarget code
I want to make it possible to customize how stops are displayed after a stop row is tapped on the Nearby Stops view controller. i.e. Make it possible to dismiss the modal Nearby Stops controller and then push onto the stack from the presenter, or just push directly onto the stack, depending on how the Nearby Stops controller was originally displayed. This is going to be a lot easier to do if I have the ability to designate that objects other than the app delegate can `-navigateToTarget:`.

* Move Deep Link code into a new `navigation` directory in OBAKit
* Move the OBANavigationTarget* code into OBAKit, and consolidate it under `navigation`
* Create the OBANavigation protocol, which specifies the `-navigateToTarget:` method implemented by `OBAApplicationDelegate` so that other objects can do the same thing