Fixes bug that caused the app to unexpectedly pick Tampa region
If the location manager is being lame and is refusing to give us a location, then we need to proactively bail on the process of picking a new region. Otherwise, Objective-C's non-clever treatment of nil will result in us unexpectedly selecting Tampa. This happens because Tampa is the closest region to lat long point (0,0).