Moves user information logging into PrivacyBroker
This ensures that, when a user says “I don’t love the app,” I can figure out what region they’re in and whether there are any other obviously diagnosable issues with their app configuration. This also makes it easier to ensure that I don’t accidentally screw something up when I add a new privacy-impacting data point to the app.

Because of how app extensions work, though, this does mean that I had to add a delegate to PrivacyBroker so that an object that isn’t going to be part of an app extension can talk to the Apptentive SDK.