Calculates walking distance to stops by absolute distance
Fixes #1027 - The 'walkable' indicator on the stop controller doesn't update location any longer

Doing 'as the crow flies' distance calculations to stops isn't as accurate as getting walking directions and taking the distance from that, but it means that this operation can be done synchronously. That means it'll be updated every time the stop data is reloaded, and it also means that those awful insertion animations will no longer be used.