Make 'status code check' an explicit feature of OBAJsonDataSource
We use a status code field in the JSON body of the OBA server's response instead of the response code header value. Previously, the determination of whether a given data source uses this behavior was based upon a pointer equality check. This change removes that potentially fraught behavior and replaces it with a property on OBAJsonDataSource that determines explicitly whether or not we should do this.