Creates Promise-focused subclass of OBAModelService
Moves regional alerts promise to the new subclass.

This change is intended to solve a few problems:

a) OBAModelService is too large and should be slimmed down. Dividing it up along the bounds of block callbacks and promises is an obvious way to slim it down.
b) OBAModelService is implemented in Objective-C, which makes it impossible to use some of the more advanced features of PromiseKit, which require Swift.