Presents the user with a captive portal browser when appropriate
Fixes #1258 - Provide a better experience when users are on a captive portal

This change will display an SFSafariViewController to the user with the URL that is failing to load when the error indicates that the user is on a captive portal wifi network. Typically, this URL should be the address that the user is being redirected to such that they can agree to terms and conditions or watch an ad, or whatever.