Creates experimental drawer UI
So, to be clear: this doesn't really do much or work well yet. THAT SAID, I strongly believe that an Apple Maps or Google Maps-style UI with a drawer to provide contextual information to the user will be the best way for our UI to evolve. I want to make sure that this is a living, breathing part of the app so that we can more easily use it, experiment with it, and ensure that we're doing the right thing. So, to that end, I have added a rudimentary drawer experience to the Map tab. You can access it by enabling debug mode (tap repeatedly on the OBA logo in the Info tab), and then go into Settings to enable the Drawer UI.

Other items of note:

* OBAMapDataLoader is now injected into OBAMapViewController (and optionally NearbyStopsViewController). This helps facilitate sharing data between these two visual components without weird, nasty gyrations.
* OBAMapDataLoader now uses a multicast delegate system. See above. Best way to facilitate this is with multicast delegation.
* OBAApplicationUI now takes an instance of OBAApplication as an initialization parameter. This is a better approach regardless of what Ux we end up using down the road.