Colors scheduled arrivals dark gray
Related to #1259 - Improve usability of 'scheduled arrival'

For scheduled arrivals—where we're just using the time table for the bus as opposed to real data—mark their data in dark gray instead of black to help try to better communicate the unreliable nature of the data the user is seeing. I don't think this is a fix for the issue, but hopefully it helps ameliorate user pain.