Embed a stripped-down stop controller in the map's callouts
We've gotten reports of map callouts not being tappable on iOS 12. I spent some time trying to figure out why they weren't working any longer, without luck. It might actually be an iOS bug. But, either way, I've been thinking about presenting more information in the callouts on the map anyway, and the work I did a few months ago to make it easy to present popovers from anywhere means that embedding a stripped-down stop controller on top of the map in a popover is pretty straightforward. I think this is a usability win.