Overhauls Analytics
* Fixes #1327 - Better Fabric analytics on accessibility usage
* Fixes #1343 - Analytics for Fare Payment apps

This change adds support for Firebase Analytics to the app, which is necessary as Fabric will be fully subsumed into Firebase in the next six months or so. Additionally, it increases the number of accessibility-related data points that we are tracking—we will now know what users' default font sizes are and whether they have high contrast mode enabled.